Parse nodes in a more intuitive order
I though order of appearance is what I wanted but looking at the
examples in the app I realized it's not. When specifying graphs in the
dot format one could add node definitions, but it had to to be done in
a less intuitive order. Consider these examples:

Old:
```
a;
b;
a -> c;
c -> b;
```

New:
```
a -> c -> d;
b;
c -> b;
```

Furthermore, the old way did not work well with the graph macro used
internally by the crate.